+++ packages/synapse/src/utils/createSynapse/index.ts
@@ -1,3 +1,3 @@
 export { createSynapse } from './createSynapse'
-export type { CreateSynapseOptions, Synapse, SynapseConfig, SynapseModule, SynapseShellConfig } from './synapse.types'
+export type { CreateSynapseOptions, Synapse, SynapseConfig, SynapseCore, SynapseModule, SynapseObjectConfig, SynapseShellConfig, SynapseWiring } from './synapse.types'
 export type { DependencyInput, ExtractPromiseType, ExtractStorageType, StorageCreatorFunction, SynapseDependency } from './types'